# swagger-movieDataset-task.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "defaultTitle",
    "description": "defaultDescription",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/api/deleteMovieActor": {
      "delete": {
        "description": "in this route /api/deleteMovieActor you can remove any relation between any movie and actor they have already relation",
        "parameters": [
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "87"
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "30"
          }
        ],
        "responses": {
          "200": {
            "description": "Auto generated using Swagger Inspector",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {}
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://localhost:8000"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:8000"
        }
      ]
    },
    "/api/addActorMovie": {
      "post": {
        "description": "in this route /api/addActorMovie you can declare a new relation is not exist between actor and movie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_addActorMovie_body"
              },
              "examples": {
                "0": {
                  "value": " {\n        \"movie_id\": 1,\n        \"actor_id\": 61\n}"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auto generated using Swagger Inspector",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {}
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://localhost:8000"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:8000"
        }
      ]
    },
...
    "/api/getOneMoviesActors": {
      "get": {
        "description": "in this route /api/getOneMoviesActors you can get any relation that we have between a movie and an actor by query their ID'S",
        "parameters": [
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "Auto generated using Swagger Inspector",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "examples": {}
              }
            }
          }
        },
        "servers": [
          {
            "url": "http://localhost:8000"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:8000"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "api_addActorMovie_body": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "integer"
          },
          "movie_id": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}